Adress failing tests
