+++ package.json
@@ -89,6 +89,7 @@
     "react-dom": "^19.2.8",
     "react-hook-form": "^7.84.0",
     "recharts": "3.10.1",
+    "server-only": "0.0.1",
     "swr": "^2.5.0",
     "url": "^0.11.3"
   },
